+++ resources/views/components/app-layout.blade.php
@@ -22,7 +22,7 @@
             @include('layouts.navigation')
 
             @isset($header)
-                <header class="bg-white border-b border-black">
+                <header class="bg-white">
                     <div class="max-w-7xl mx-auto py-6 px-4 sm:px-6 lg:px-8">
                         {{ $header }}
                     </div>